+++ Assetra.WPF/Infrastructure/ReportsServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 using Assetra.Application.Reports;
 using Assetra.Application.Reports.Statements;
+using Assetra.Core.Interfaces;
+using Assetra.Core.Interfaces.Analysis;
 using Assetra.Core.Interfaces.Reports;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -11,7 +13,12 @@ internal static class ReportsServiceCollectionExtensions
     {
         services.AddSingleton<Assetra.Application.Reports.Services.MonthEndReportService>();
         services.AddSingleton<IIncomeStatementService, IncomeStatementService>();
-        services.AddSingleton<IBalanceSheetService, BalanceSheetService>();
+        services.AddSingleton<IBalanceSheetService>(sp => new BalanceSheetService(
+            sp.GetRequiredService<IAssetRepository>(),
+            sp.GetRequiredService<ITradeRepository>(),
+            sp.GetService<IPortfolioSnapshotRepository>(),
+            sp.GetService<IMultiCurrencyValuationService>(),
+            sp.GetService<IAppSettingsService>()?.Current.BaseCurrency));
         services.AddSingleton<ICashFlowStatementService, CashFlowStatementService>();
         services.AddSingleton<IReportExportService, ReportExportService>();
         services.AddSingleton<Assetra.WPF.Features.Reports.ReportsViewModel>();